+++ resources/command.go
@@ -1,4 +1,4 @@
-package utils
+package resources
 
 import (
 	"bytes"
